+++ packages/react/src/CuriBase.tsx
@@ -3,9 +3,11 @@ import PropTypes from "prop-types";
 import { CuriRouter, Response, Navigation } from "@curi/core";
 import { CuriContext, CuriProps } from "./interface";
 
+export type CuriRenderFn = (props: CuriProps) => React.ReactElement<any>;
+
 export interface CuriBaseProps {
   router: CuriRouter;
-  render: (props: CuriProps) => React.ReactElement<any>;
+  render: CuriRenderFn;
   response: Response;
   navigation: Navigation;
 }